changed perlin 2D 2/28/25

+++ src/main/java/com/tardisgallifrey/PerlinNoise/Point.java
@@ -1,4 +1,4 @@
 package com.tardisgallifrey.PerlinNoise;
 
-public record Point(int y, int x, int height) {
+public record Point(int y, int x, double height) {
 }
